fix compiler warning: subunit header file not found
The code is using ENABLE_SUBUNIT to determine if subunit is available.
HAVE_SUBUNIT_CHILD_H is never defined. When subunit is used, its
header files are never found, causing a compiler warning. This commit
replaces the #if protecting the header to be consistent with the
rest of the code.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@660 64e312b2-a51f-0410-8e61-82d0ca0eb02a