modify test_check_nofork.sh to compare output based on HAVE_FORK
When run in Windows, the output differs slightly from when it is run
under *NIX due to line endings. *NIX uses \n, whereas Windows
uses \r\n. To account for this, printf is used to generate the
Windows expected string.

One other change is not using stderr for the test. Wine has an issue
with the tmpfile() call, and will output:
   fixme:msvcrt:MSVCRT__sopen_s : pmode 0x0033 ignored
to stderr when tmpfile() is called. Besides, if the test actually
does segfault in the future (which is the reason for grabbing the stderr),
the result will not be to print out the test results.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@781 64e312b2-a51f-0410-8e61-82d0ca0eb02a