disable checked signal tests if fork is unavailable
As mentioned in a previous commit, checked fixtures are disabled
if fork is unavailable, as they are unable to provide the guarantees
that they advertise without fork.

In addition, some line numbers are updated. This is because some tests
are looking through error messages, and the line number of the
messages is important.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@796 64e312b2-a51f-0410-8e61-82d0ca0eb02a