if tmpfile() fails, remember the filename and delete it later
One feature of tmpfile() is to unlink the file, causing it to be
deleted after it is closed. Unlinking an open file fails in Windows.
To prevent the case where the temp directory just fills up with
files, the created file name is remembered, and deleted when
it is closed.

Certainly, if a unit test fails before it deletes the file,
the temp directory will still fill up. However, this is better
than not deleting the files.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@814 64e312b2-a51f-0410-8e61-82d0ca0eb02a