If clocks are not defined, set to (hopefully) invalid value
Likely, if clock_gettime() is implemented on a system, the first
valid clock will be set to '0'. If not all clock are defined,
by setting our fake definitions to '0', we may accidental use a
valid clock when we did not mean to. Setting the clocks to '-1'
in hopes to avoid this.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@819 64e312b2-a51f-0410-8e61-82d0ca0eb02a