Removing const identifier from List data structure
The const identifier on the data field of List would imply that
nothing is ever done to modify or destroy the data. However, the
data can be accessed later and modified or freed, as was being done
in places. By removing the const identifiers, the implied safety
of the stored values (which is being violated anyway) is not
advertised in the API.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@1063 64e312b2-a51f-0410-8e61-82d0ca0eb02a