Run indent on source
The indent program was run against the source, to enforce a consistent
layout and spacing. The following arguments were used:

--blank-lines-after-declarations --blank-lines-after-procedures
--brace-indent0 --case-indentation4 --case-brace-indentation0
--no-space-after-function-call-names --no-space-after-casts
--declaration-indentation1 --no-blank-lines-after-commas
--braces-after-struct-decl-line --braces-after-func-def-line
--continue-at-parentheses --preprocessor-indentation0  --indent-level4
--no-tabs --dont-break-procedure-type --no-space-after-parentheses
--no-space-after-for --no-space-after-if --no-space-after-while

Note that this was not run againt the test files for two reasons:
1) some of the tests depend on the failures occuring on specific lines,
and I did not want to go through and fix that for each test.
2) indent does not know what to do with END_TEST, and thinks it is a
type modifier for the function that follows.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@1081 64e312b2-a51f-0410-8e61-82d0ca0eb02a