upon test failure, invoke _exit() instead of exit()
exit() will call atexit handlers, which may try to
clean up things or wait for things to get cleaned up,
which we don't want or need. We just want to stop
and let the parent know about the failure as quickly
as possible in case fork() is used.

Fixes timeouts on assert failures in checks where
an exit handler waits for things to stop, but they
don't stop because they haven't been shut down,
and they haven't been shut down because there's no
simple way to do so on failures.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@1128 64e312b2-a51f-0410-8e61-82d0ca0eb02a