autotools: remove AIX fix for broken xlc v6 compiler
The issue with the xlc v6 compiler on AIX seems to be a compiler
bug which is resolved in future releases of the compiler.

This fix, as it turns out, did not properly detect the bug
or resolve the issue. Unless it becomes really important,
the attempt to get around the bug is removed from Check.

git-svn-id: svn://svn.code.sf.net/p/check/code/trunk@1173 64e312b2-a51f-0410-8e61-82d0ca0eb02a